Move repo introspection as static type of Repository interface.
This is the first step of refactoring for creating non-aggregate-root repositories.
