--- core-java/src/main/java/org/spine3/util/ListFilters.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2015, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.spine3.util;
-
-import com.google.common.base.Predicate;
-import com.google.common.collect.FluentIterable;
-import com.google.common.collect.ImmutableList;
-
-/**
- * Provides utility for filtering lists with predicates.
- *
- * @author Mikhail Mikhayloc
- */
-@SuppressWarnings("UtilityClass")
-public class ListFilters {
-
-    private ListFilters() {
-    }
-
-    /**
-     * Filters list with a predicate.
-     *
-     * @param list      a list to filter
-     * @param predicate a predicate to filter list
-     * @param <T>       list object type
-     * @return immutable list
-     */
-    public static <T> ImmutableList<T> filter(Iterable<T> list, Predicate<T> predicate) {
-        return FluentIterable.from(list).filter(predicate).toList();
-    }
-}
